requirements.txt
opencv-python
numpy
pillow
asyncio
websockets

